add agent related messages
