+++ langserve/serialization.py
@@ -8,6 +8,7 @@ from typing import Any, Union
 
 from langchain.prompts.base import StringPromptValue
 from langchain.prompts.chat import ChatPromptValueConcrete
+from langchain.schema.agent import AgentAction, AgentActionMessageLog, AgentFinish
 from langchain.schema.document import Document
 from langchain.schema.messages import (
     AIMessage,
@@ -50,6 +51,9 @@ class WellKnownLCObject(BaseModel):
         AIMessageChunk,
         StringPromptValue,
         ChatPromptValueConcrete,
+        AgentAction,
+        AgentFinish,
+        AgentActionMessageLog,
     ]
